+++ packages/builder-tools/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 6.1.0-dev.4 (2026-05-26)
+
+### 🩹 Fixes
+
+- **builder-tools:** include @powerhousedao/connect in connect-utils linkedRoots ([2865fa07f](https://github.com/powerhouse-inc/powerhouse/commit/2865fa07f))
+
+### ❤️ Thank You
+
+- acaldas
+
+## 6.1.0-dev.3 (2026-05-25)
+
+This was a version bump only for @powerhousedao/builder-tools to align it with other projects, there were no code changes.
+
+## 6.1.0-dev.2 (2026-05-24)
+
+This was a version bump only for @powerhousedao/builder-tools to align it with other projects, there were no code changes.
+
+## 6.1.0-dev.1 (2026-05-23)
+
+This was a version bump only for @powerhousedao/builder-tools to align it with other projects, there were no code changes.
+
 ## 6.1.0-dev.0 (2026-05-22)
 
 ### 🩹 Fixes